moved API out of src
