+++ api/routes.php
@@ -1,0 +1,7 @@
+<?php
+
+use HomeCEU\DTS\Api;
+
+return [
+  new Api\Route('post','/docdata', Api\DocData\DocDataAdd::class)
+];
