--- config/routes.php
@@ -1,7 +1,0 @@
-<?php
-
-use HomeCEU\DTS\Api;
-
-return [
-  new Api\Route('post','/docdata', Api\PostDocData::class)
-];
